+++ StripeIdentity/StripeIdentity/Source/API Bindings/Models/VerificationPageData/VerificationPageData.swift
@@ -22,6 +22,21 @@ extension StripeAPI {
         let status: Status
         /// If true, the associated VerificationSession has been submitted for processing.
         let submitted: Bool
+
+        /// If true, the associated VerificationSession has been closed and can no longer be modified.
+        /// After submitting, closed might be false if needs to fallback from phone verification to document verification.
+        let closed: Bool
     }
 
+}
+
+extension StripeAPI.VerificationPageData {
+    /// When submitted but is not closed and there is still missing requirements, need to fallback.
+    func needsFallback() -> Bool {
+        return submitted && !closed && !requirements.missing.isEmpty
+    }
+
+    func submittedAndClosed() -> Bool {
+        return submitted && closed
+    }
 }