+++ StripeIdentity/StripeIdentityTests/Helpers/VerificationSheetControllerMock.swift
@@ -59,6 +59,12 @@ final class VerificationSheetControllerMock: VerificationSheetControllerProtocol
 
     var completeOption: CompleteOptionView.CompleteOption?
 
+    var generatePhonOtpSuccessCallback: ((StripeCore.StripeAPI.VerificationPageData) -> Void)?
+    var cannotVerifyPhoneOtpCalled: Bool = false
+
+    var saveOtpAndMaybeTransitionCompletion: (() -> Void)?
+    var saveOtpAndMaybeTransitionInvalidOtp: (() -> Void)?
+
     init(
         apiClient: IdentityAPIClient = IdentityAPIClientTestMock(),
         flowController: VerificationSheetFlowControllerProtocol =
@@ -139,12 +145,26 @@ final class VerificationSheetControllerMock: VerificationSheetControllerProtocol
         }
     }
 
+    func saveOtpAndMaybeTransition(from fromScreen: StripeIdentity.IdentityAnalyticsClient.ScreenName, otp otpValue: String, completion: @escaping () -> Void, invalidOtp: @escaping () -> Void) {
+        saveOtpAndMaybeTransitionCompletion = completion
+        saveOtpAndMaybeTransitionInvalidOtp = invalidOtp
+
+    }
+
     func verifyAndTransition(simulateDelay: Bool) {
         completeOption = simulateDelay ? .successAsync : .success
     }
 
     func unverifyAndTransition(simulateDelay: Bool) {
         completeOption = simulateDelay ? .failureAsync : .failure
+    }
+
+    func generatePhoneOtp(using successCallback: @escaping (StripeCore.StripeAPI.VerificationPageData) -> Void) {
+        generatePhonOtpSuccessCallback = successCallback
+    }
+
+    func sendCannotVerifyPhoneOtpAndTransition(completion: @escaping () -> Void) {
+        self.cannotVerifyPhoneOtpCalled = true
     }
 
     func transitionToCountryNotListed(missingType: StripeIdentity.IndividualFormElement.MissingType) {